Refactor of customer list
